+++ kolezza_dash_board/src/app/hooks/useGetChildren.ts
@@ -1,43 +1,39 @@
 import { useState, useEffect } from 'react';
-import { fetchChildren } from '@/app/utils/fetchChildren'; // Assuming this utility exists
-import { FetchChildrenResponse } from '@/app/utils/types';
+import { fetchAllChildren } from '@/app/utils/fetchChildren';
+import { FetchChildrenResponse, Child } from '@/app/utils/types';
 
-export function useChildren() {
-    const [state, setState] = useState({
-        activePatients: 0,
-        inactivePatients: 0,
-        loading: true,
-        error: null as string | null,
-    });
+export const useChildren = () => {
+  const [activePatients, setActivePatients] = useState<number>(0);
+  const [inactivePatients, setInactivePatients] = useState<number>(0);
+  const [loading, setLoading] = useState<boolean>(true);
+  const [error, setError] = useState<string | null>(null);
 
-    const { activePatients, inactivePatients, loading, error } = state;
-
+  useEffect(() => {
     const getChildren = async () => {
-        setState(prev => ({ ...prev, loading: true, error: null }));
-        try {
-            const response: FetchChildrenResponse = await fetchChildren(); 
-            const children = response.child || []; 
-            const active = children.filter(child => !child.is_deleted).length;
-            const inactive = children.filter(child => child.is_deleted).length;
+      setLoading(true);
+      setError(null);
+      try {
+        console.log('Fetching children from API...');
+        const response: FetchChildrenResponse = await fetchAllChildren();
+        console.log('API response:', response);
 
-            setState({
-                activePatients: active,
-                inactivePatients: inactive,
-                loading: false,
-                error: null,
-            });
-        } catch (err) {
-            setState(prev => ({
-                ...prev,
-                loading: false,
-                error: (err as Error).message,
-            }));
-        }
+        const children: Child [] = response?.child ?? [];
+
+        const active = children.filter(child => !child.is_deleted).length;
+        const inactive = children.filter(child => child.is_deleted).length;
+
+        setActivePatients(active);
+        setInactivePatients(inactive);
+      } catch (err) {
+        console.error('Error fetching children:', err);
+        setError(err instanceof Error ? err.message : 'An unexpected error occurred.');
+      } finally {
+        setLoading(false);
+      }
     };
 
-    useEffect(() => {
-        getChildren();
-    }, []); 
+    getChildren();
+  }, []);
 
-    return { activePatients, inactivePatients, loading, error };
-}
+  return { activePatients, inactivePatients, loading, error };
+};
